Cleanup for the demos. Move benchmarks to own directory.
